Area_server/requirements.txt
Django==3.2.20
djangorestframework==3.14.0
psycopg2-binary==2.9.7
google-auth-httplib2
google-auth-oauthlib
google-auth==2.23.4
django-background-tasks==1.2.5
google-api-python-client==2.102.0
django-allauth==0.57.0
celery==5.3.4
kombu
redis
requests==2.31.0
spotipy
python-dotenv==1.0.0
django-cors-headers==4.3.0